pyproject.toml

[project]
name = "chameleonflow"
version = "0.1.0"
description = "Adaptive transport system prototype with client, server, and ML components"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "aiohttp>=3.13.5",
    "aioquic>=1.3.0",
    "aiortc>=1.14.0",
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "lightgbm>=4.6.0",
    "numpy>=2.4.4",
    "onnx>=1.21.0",
    "onnxruntime>=1.24.4",
    "pandas>=3.0.2",
    "prometheus-client>=0.24.1",
    "pydantic-settings>=2.13.1",
    "pynacl>=1.6.2",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "sqlalchemy>=2.0.49",
    "structlog>=25.5.0",
    "typer>=0.24.1",
    "uvicorn>=0.44.0",
    "vowpalwabbit>=9.11.2",
]

[project.optional-dependencies]
train = [
    "onnxmltools>=1.14.0",
    "pyarrow>=21.0.0",
    "torch>=2.11.0",
]

[tool.uv.sources]
torch = { index = "pytorch-gpu" }

[[tool.uv.index]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.20.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.9",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
